requirements-pytest.txt
django-bootstrap3
django-countries
django-tabbed-admin
python-slugify
tox
django-exclusivebooleanfield
djangorestframework>=3.5.4
django-swapfield
django-filter
unicodecsv